Update xrpl-cheat-sheet.json
www.dogoninu.com Is a project on the XRPL by the students at Uhuru cultural arts institute/Training Center. We want to build a Computer lab that we will name Project R.A.Y for Ripple Africa Youth which will be a technology lab to support and improve our facility and the local youth that attend our programs so they can benefit from the program as we have.